fix: Move cloud loadbalancer command under network
