+++ internal/cmd/cloud_project.go
@@ -51,7 +51,6 @@ func init() {
 	initContainerRegistryCommand(cloudCmd)
 	initCloudDatabaseCommand(cloudCmd)
 	initInstanceCommand(cloudCmd)
-	initCloudLoadbalancerCommand(cloudCmd)
 	initCloudNetworkCommand(cloudCmd)
 	initCloudOperationCommand(cloudCmd)
 	initCloudQuotaCommand(cloudCmd)
